refactored NavHost and mainscreen
