+++ app/src/main/AndroidManifest.xml
@@ -52,6 +52,16 @@
                 <data android:host="app.stage.rarime.com" />
                 <data android:pathPattern="/r/.*" />
             </intent-filter>
+            <intent-filter android:autoVerify="true">
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data android:scheme="http" />
+                <data android:host="app.stage.rarime.com" />
+                <data android:pathPattern="/r/.*" />
+            </intent-filter>
         </activity>
 
         <activity-alias
@@ -85,6 +95,7 @@
                 <category android:name="android.intent.category.BROWSABLE" />
 
                 <data android:scheme="https" />
+                <data android:scheme="http" />
                 <data android:host="app.stage.rarime.com" />
                 <data android:pathPattern="/r/.*" />
             </intent-filter>
@@ -121,6 +132,7 @@
                 <category android:name="android.intent.category.BROWSABLE" />
 
                 <data android:scheme="https" />
+                <data android:scheme="http" />
                 <data android:host="app.stage.rarime.com" />
                 <data android:pathPattern="/r/.*" />
             </intent-filter>
@@ -157,6 +169,7 @@
                 <category android:name="android.intent.category.BROWSABLE" />
 
                 <data android:scheme="https" />
+                <data android:scheme="http" />
                 <data android:host="app.stage.rarime.com" />
                 <data android:pathPattern="/r/.*" />
             </intent-filter>
