+++ app/src/main/java/com/rarilabs/rarime/store/SecureSharedPrefsManager.kt
@@ -17,9 +17,6 @@ interface SecureSharedPrefsManager {
 
     fun readPrivateKey(): String?
 
-    fun readIsIntroFinished(): Boolean
-    fun saveIsIntroFinished(isFinished: Boolean)
-
     fun readPasscodeState(): SecurityCheckState
     fun savePasscodeState(state: SecurityCheckState)
